requirements.txt
Django==3.0.7
django-admin-sortable2==0.7.6
django-tinymce==3.0.2
Pillow==7.2.0
python-dotenv==0.14.0
requests==2.24.0
django-debug-toolbar==2.2
environs==8.0.0